+++ chat/src/jvmMain/kotlin/data/Message.kt
@@ -3,8 +3,8 @@ package data
 import java.time.LocalDateTime
 
 data class Message(
-  val sender: Long,
-  val receiver: Long,
+  val sender: Int,
+  val receiver: Int,
   val message: String,
   val time: LocalDateTime
 )
